Move all config classes from the Configuration project to Services and delete the old project.
