--- src/Configuration/Properties/AssemblyInfo.cs
@@ -1,24 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-#if DEBUG
-[assembly: AssemblyTitle("SenseNet.Configuration (Debug)")]
-#else
-[assembly: AssemblyTitle("SenseNet.Configuration (Release)")]
-#endif
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Sense/Net Inc.")]
-[assembly: AssemblyCopyright("Copyright © Sense/Net Inc.")]
-[assembly: AssemblyProduct("sensenet")]
-[assembly: AssemblyTrademark("Sense/Net Inc.")]
-[assembly: AssemblyCulture("")]
-
-[assembly: AssemblyVersion("7.5.0.2")]
-[assembly: AssemblyFileVersion("7.5.0.2")]
-[assembly: AssemblyInformationalVersion("7.5.0.2")]
-
-[assembly: ComVisible(false)]
-[assembly: Guid("dfbdb163-d9bb-481c-b3fd-e9eb0e37d27d")]
-
