+++ src/SenseNet.sln
@@ -12,8 +12,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SenseNet.Storage", "Storage
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SenseNet.Services", "Services\SenseNet.Services.csproj", "{B72529C8-FEB1-49F5-B08B-56055B58F296}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SenseNet.Configuration", "Configuration\SenseNet.Configuration.csproj", "{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SnAdminRuntime", "Tools\SnAdminRuntime\SnAdminRuntime.csproj", "{891B1FD9-CA73-4129-8D1D-D6663F01866C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SenseNet.Common", "Common\SenseNet.Common.csproj", "{A453E920-29C0-45CD-984C-0D8E3631B1E3}"
@@ -99,18 +97,6 @@ Global
 		{B72529C8-FEB1-49F5-B08B-56055B58F296}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{B72529C8-FEB1-49F5-B08B-56055B58F296}.Release|x64.ActiveCfg = Release|Any CPU
 		{B72529C8-FEB1-49F5-B08B-56055B58F296}.Release|x86.ActiveCfg = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|x64.ActiveCfg = Release|Any CPU
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161}.Release|x86.ActiveCfg = Release|Any CPU
 		{891B1FD9-CA73-4129-8D1D-D6663F01866C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{891B1FD9-CA73-4129-8D1D-D6663F01866C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{891B1FD9-CA73-4129-8D1D-D6663F01866C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -339,7 +325,6 @@ Global
 		{786E6165-CA02-45A9-BF58-207A45D7D6DF} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
 		{5DB4DDBA-81F6-4D81-943A-18F3178B3355} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
 		{B72529C8-FEB1-49F5-B08B-56055B58F296} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
-		{A19F707C-B1C3-45C7-AE38-E7B7F30C1161} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
 		{891B1FD9-CA73-4129-8D1D-D6663F01866C} = {875EF569-4684-473D-A2D4-A35B20B4A07C}
 		{A453E920-29C0-45CD-984C-0D8E3631B1E3} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
 		{78FEBD73-8BC4-45B9-9B31-D512E1AFAD60} = {2997D17C-A736-43E5-B3DD-11D11AC7DF17}
