+++ src/Tests/SenseNet.ContentRepository.Tests/SenseNet.ContentRepository.Tests.csproj
@@ -143,10 +143,6 @@
       <Project>{a453e920-29c0-45cd-984c-0d8e3631b1e3}</Project>
       <Name>SenseNet.Common</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\Configuration\SenseNet.Configuration.csproj">
-      <Project>{a19f707c-b1c3-45c7-ae38-e7b7f30c1161}</Project>
-      <Name>SenseNet.Configuration</Name>
-    </ProjectReference>
     <ProjectReference Include="..\..\ContentRepository\SenseNet.ContentRepository.csproj">
       <Project>{786e6165-ca02-45a9-bf58-207a45d7d6df}</Project>
       <Name>SenseNet.ContentRepository</Name>
